Split outbound test modules into files
Several outbound modules include embedded stack tests. This change moves
these submodules into distinct files.
